+++ test/test_main.py
@@ -16,15 +16,12 @@ def make_request_obj(status):
 
 class TestSendApiRequest(unittest.TestCase):
 
-    def setUp(self):
-        self.base_url = "https://reqres.in/api/users/2"
-        self.method = "GET"
-        self.params = {"param1": "value1", "param2": "value2"}
-        self.data = {"key1": "value1", "key2": "value2"}
-        self.headers = {"Content-Type": "application/json"}
-        self.handler = APIRequestHandler(headers=self.headers)
-
     def test_happy_path(self):
+        base_url = "https://reqres.in/api/users/2"
+        method = "GET"
+        params = {"param1": "value1", "param2": "value2"}
+        headers = {"Content-Type": "application/json"}
+        handler = APIRequestHandler(headers=headers)
         validate_return = True
         expected = {
             "data": {
@@ -39,74 +36,120 @@ class TestSendApiRequest(unittest.TestCase):
                 "text": "To keep ReqRes free, contributions towards server costs are appreciated!"
             }
         }
-        with patch("raspberryrequest.main.valid_status") as mock_valid_status:
-            mock_valid_status.return_value = validate_return
-            response = self.handler.send_api_request(
-                self.base_url, self.method, self.params, self.headers)
+        with patch("raspberryrequest.main.validate_status") as mock_validate_status:
+            mock_validate_status.return_value = validate_return
+            response = handler.send_api_request(
+                base_url, method, params, headers)
         self.assertDictEqual(expected, response)
 
     def test_timeout_error(self):
+        base_url = "https://reqres.in/api/users/2"
+        method = "GET"
+        params = {"param1": "value1", "param2": "value2"}
+        headers = {"Content-Type": "application/json"}
+        handler = APIRequestHandler(headers=headers)
         with patch("requests.Session.send") as mock_send:
             mock_send.side_effect = requests.exceptions.Timeout
             with self.assertRaises(MaxRetryError):
-                self.handler.send_api_request(
-                    self.base_url, self.method, self.params, self.headers)
-            attempts = self.handler.calls
+                handler.send_api_request(
+                    base_url, method, params, headers)
+            attempts = handler.calls
             self.assertEqual(attempts, 3)
 
     def test_HTTPError(self):
+        base_url = "https://reqres.in/api/users/2"
+        method = "GET"
+        params = {"param1": "value1", "param2": "value2"}
+        headers = {"Content-Type": "application/json"}
+        handler = APIRequestHandler(headers=headers)
         with patch("requests.Session.send") as mock_send:
             mock_send.side_effect = requests.exceptions.HTTPError
             with self.assertRaises(MaxRetryError):
-                self.handler.send_api_request(
-                    self.base_url, self.method, self.params, self.headers)
-            attempts = self.handler.calls
+                handler.send_api_request(
+                    base_url, method, params, headers)
+            attempts = handler.calls
             self.assertEqual(attempts, 3)
 
     def test_NonRetryableStatusCodeError(self):
-        self.base_url = "https://reqres.in/api/users/23"
-        result = self.handler.send_api_request(
-            self.base_url, self.method, self.params, self.headers)
+        base_url = "https://reqres.in/api/users/23"
+        method = "GET"
+        params = {"param1": "value1", "param2": "value2"}
+        headers = {"Content-Type": "application/json"}
+        handler = APIRequestHandler(headers=headers)
+        result = handler.send_api_request(
+            base_url, method, params, headers)
         self.assertIsNone(result)
 
     def test_FatalStatusCodeError(self):
+        base_url = "https://reqres.in/api/users/2"
+        method = "GET"
+        params = {"param1": "value1", "param2": "value2"}
+        headers = {"Content-Type": "application/json"}
+        handler = APIRequestHandler(headers=headers)
         response = requests.Response()
         response.status_code = 403
         response.reason = "test"
         with patch("requests.Session.send") as mock_send:
             mock_send.return_value = response
             with self.assertRaises(FatalStatusCodeError):
-                self.handler.send_api_request(
-                    self.base_url, self.method, self.params, self.headers)
+                handler.send_api_request(
+                    base_url, method, params, headers)
 
     def test_close_session(self):
-        self.handler.close_session()
+        headers = {"Content-Type": "application/json"}
+        handler = APIRequestHandler(headers=headers)
+        handler.session_data.VALID = 1
+        handler.close_session()
+
+        valid = handler.session_data.VALID
+        retryable = handler.session_data.RETRYABLE
+        nonretryable = handler.session_data.NONRETRYABLE
+        fatal = handler.session_data.FATAL
+        paid = handler.session_data.PAID
+        unpaid = handler.session_data.UNPAID
+        self.assertEqual(valid, 0)
+        self.assertEqual(retryable, 0)
+        self.assertEqual(nonretryable, 0)
+        self.assertEqual(fatal, 0)
+        self.assertEqual(paid, 0)
+        self.assertEqual(unpaid, 0)
 
     def test_add_status_code(self):
+        headers = {"Content-Type": "application/json"}
+        handler = APIRequestHandler(headers=headers)
         new_status_code = 260
         expected_list = [200, 201, 260]
-        self.handler.add_status_code("VALID", new_status_code)
-        list = self.handler.status_codes.VALID
+        handler.add_status_code("VALID", new_status_code)
+        list = handler.status_codes.VALID
 
         self.assertListEqual(expected_list, list)
 
     def test_remove_status_code(self):
+        headers = {"Content-Type": "application/json"}
+        handler = APIRequestHandler(headers=headers)
         removed_status_code = 260
         expected_list = [200, 201]
-        self.handler.remove_status_code("VALID", removed_status_code)
-        list = self.handler.status_codes.VALID
+        handler.remove_status_code("VALID", removed_status_code)
+        list = handler.status_codes.VALID
 
         self.assertListEqual(expected_list, list)
 
     def test_get_status_codes(self):
-        status_codes = self.handler.get_status_codes()
+        headers = {"Content-Type": "application/json"}
+        handler = APIRequestHandler(headers=headers)
+        status_codes = handler.get_status_codes()
         self.assertEqual(status_codes, handler.status_codes)
 
     def test_print_status_codes(self):
-        self.handler.print_status_codes()
+        headers = {"Content-Type": "application/json"}
+        handler = APIRequestHandler(headers=headers)
+        handler.print_status_codes()
 
-    def tearDown(self):
-        self.handler = APIRequestHandler(headers=self.headers)
+    def test_get_session_codes(self):
+        headers = {"Content-Type": "application/json"}
+        handler = APIRequestHandler(headers=headers)
+        session_data = handler.get_session_data()
+        self.assertIsInstance(session_data, dict)
 
 
 if __name__ == '__main__':